Move GetterOps stuff out of Utils.hs (+ remove Lift)
